% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zarr.R
\name{read_zarr}
\alias{read_zarr}
\alias{write_zarr}
\title{Zarr I/O}
\usage{
read_zarr(file, zarr_version = NULL, zarr_path = NULL)

write_zarr(x, file, zarr_version = NULL, zarr_path = NULL, ...)
}
\arguments{
\item{file}{path to the file to be read/written}

\item{zarr_version}{The zarr protocol version to use when saving. The
default value of None will attempt to infer the version from store if
possible, otherwise it will fall back to 2.}

\item{zarr_path}{The path within the group where the arrays will be saved.}

\item{x}{object to save into the \code{zarr} file}

\item{...}{not currently used.}
}
\value{
The output from \code{zarr.load}
}
\description{
Zarr I/O
}
\examples{
file = system.file("extdata/example.zip", package = "zarrr")
out = read_zarr(file)
}
