% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_zarr.R
\name{install_zarr}
\alias{install_zarr}
\alias{have_zarr}
\alias{zarr_version}
\title{Install the \code{zarr} Python Module}
\usage{
install_zarr(..., packages = "zarr")

have_zarr()

zarr_version()
}
\arguments{
\item{...}{Additional arguments to pass to \code{\link[reticulate:py_install]{reticulate::py_install()}}}

\item{packages}{additional packages to install.  If \code{zarr} is missing,
it will be added}
}
\value{
Output of \link[reticulate:py_install]{reticulate::py_install}
}
\description{
Install the \code{zarr} Python Module
}
\examples{
if (have_zarr()) {
   zarr_version()
}

}
